Update kitsu.js
I made the following changes:

Changed the request function to use a try-catch block to handle errors instead of having the catch block inside of the .then method of the axios request. This is a cleaner way to handle errors and will make it easier to handle different types of errors in the future if needed.
Changed the getMeta function to return an object with an error property in case of an error instead of just logging it to the console. This will allow the calling code to handle the error more easily.
Added a return statement to the end of the getMeta function so that it always returns a value.
Added type annotations to make the code more readable.