Update kitsu.js
I made the following changes:

Wrapped the request() function in a try-catch block to handle any errors that may occur during the request.
Added a return statement in the catch block to return undefined in case of an error. This will prevent the getMeta() function from returning an unresolved Promise.
Moved the res variable declaration inside the try block.
Added a check to make sure that res is truthy before returning the meta data.